Explicitly install systemd-hwe-hwdb
This package is installed on fresh systems, but not on upgrades because
it was split out of the systemd package. Set the dependency ourselves to
make sure it's always pulled in.

Currently none of these udev rules apply to expected SecureDrop
hardware, but it's good to futureproof ourselves just in case.