noble-migration: Run migration check first
We currently run `pending_updates()` and then run `migration_check()`,
and the initial thinking was that we should update everything to latest
before checking stuff, which is reasonable, except at some point I
refactored it to also disable unattended-updates during
`pending_updates()` to take advantage of the fact that
`pending_updates()` does a reboot.

The only problem is that we expect the `migration_check()` to somewhat
be fallible. But we shouldn't end up in a state where it fails and then
now updates are fully stopped, needing manual intervention.

So the fix is to simply reorder them, which is relatively trivial,
though the manual index in the semiautomated playbook also needs
updating.

It's theoretically possible that `pending_updates()` brings in something
that would have caused `migration_check()` to fail, but at this point I
think it's pretty unlikely given that the fully automated migration is
only going to be kicked off immediately after pending updates are
applied.

Fixes #7514.